2023/8/13

+++ src/Problems/CodeForces/CF1A.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int n, m, a;
+    unsigned long long n, m, a;
     cin >> n >> m >> a;
     if (n % a == 0)
         n = n / a;
